+++ src/components/dashboard/Dashboard.css
@@ -18,81 +18,6 @@
     width: 40%;
 }
 
-/* Add Budget */
-.startBudget {
-    display: flex;
-    justify-content: flex-start;
-    margin-top: 10px;
-}
-
-.startBudgetList {
-    background-color: transparent;
-    width: 130px;
-    text-decoration: none;
-    list-style: none;
-    padding: 0;
-    margin: 0;
-    display: flex;
-    flex-wrap: wrap;
-    flex-direction: row;
-    justify-content: flex-start;
-    align-items: center;
-    vertical-align: middle;
-    padding: 4px 24px 4px 26px;
-}
-
-    .addBudgetBtn {
-        font-family: 'Montserrat', Arial, Helvetica, sans-serif;
-        font-size: 12px;
-        font-weight: bold;
-        height: 28px;
-        width: 120px;
-        color: #3863a0;
-        background-color: #61e5b4;
-        border: 1px solid #8fec9b;
-        border-radius: 2px;
-        box-shadow: 1px 2px 8px #d4d4d4;
-        padding: 1px 5px 2px;
-        margin: 0 0 4px;
-    }
-
-    .addBudgetBtn:hover{
-        color: #fff;
-        border-color: #5cbfed;
-        background-color: #5cbfed;
-        padding-bottom: 0px;
-    }
-
-    .selectBudget {
-        background-color: transparent;
-        padding-bottom: 1px;
-        margin-bottom: 2px;
-    }
-
-    .selectBudgetDropdown {
-        font-family: 'Montserrat', Arial, Helvetica, sans-serif;
-        font-size: 12px;
-        font-weight: 500;
-        height: 28px;
-        width: 200px;
-        background-color: #eef1f5;
-        /* border: 1px solid #9fafc1; */
-        border: none;
-        border-radius: 2px;
-        /* box-shadow: 1px 2px 8px #d4d4d4; */
-        padding: 0 0 2px 5px;
-    }
-
-    /* .selectBudgetDropdown:hover {
-        color: #FFF;
-        background-color: #9fafc1;
-        border: none;
-    } */
-
-    .bold {
-        font-weight: 500;
-        font-style: italic;
-    }
